container: use uint64 for representing memory on container config
Closes #119.

+++ container.go
@@ -158,8 +158,8 @@ type Config struct {
 	Hostname        string
 	Domainname      string
 	User            string
-	Memory          float64
-	MemorySwap      float64
+	Memory          uint64
+	MemorySwap      uint64
 	CpuShares       int64
 	AttachStdin     bool
 	AttachStdout    bool
